Fix logic from OptionGroup service and develop new Option Item Service

+++ merchant-service/src/main/java/com/fo_product/merchant_service/dtos/responses/OptionItemResponse.java
@@ -2,11 +2,13 @@ package com.fo_product.merchant_service.dtos.responses;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record OptionItemResponse(
         Long id,
         String name,
-        long priceAdjustment,
+        BigDecimal priceAdjustment,
         boolean isAvailable,
         String optionGroupName
 ) {
